Move library logic into lib.rs
