+++ src/manifest.rs
@@ -4,31 +4,31 @@ use serde_with::serde_as;
 use crate::chunk_id::FIoChunkIdRaw;
 
 #[derive(Debug, Serialize, Deserialize)]
-pub(crate) struct PackageStoreManifest {
-    pub(crate) oplog: OpLog,
+pub struct PackageStoreManifest {
+    pub oplog: OpLog,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
-pub(crate) struct OpLog {
-    pub(crate) entries: Vec<Op>,
+pub struct OpLog {
+    pub entries: Vec<Op>,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
-pub(crate) struct Op {
-    pub(crate) packagestoreentry: PackageStoreEntry,
-    pub(crate) packagedata: Vec<ChunkData>,
-    pub(crate) bulkdata: Vec<ChunkData>,
+pub struct Op {
+    pub packagestoreentry: PackageStoreEntry,
+    pub packagedata: Vec<ChunkData>,
+    pub bulkdata: Vec<ChunkData>,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
-pub(crate) struct PackageStoreEntry {
-    pub(crate) packagename: String,
+pub struct PackageStoreEntry {
+    pub packagename: String,
 }
 
 #[serde_as]
 #[derive(Debug, Serialize, Deserialize)]
-pub(crate) struct ChunkData {
+pub struct ChunkData {
     #[serde_as(as = "serde_with::hex::Hex")]
-    pub(crate) id: FIoChunkIdRaw,
-    pub(crate) filename: String,
+    pub id: FIoChunkIdRaw,
+    pub filename: String,
 }
